refactor(backend): tag bookmark contoroller 내부 조회 로직 bookmark controller로 이전
