+++ backend/taggle-core/src/main/java/kr/taggle/bookmark/service/BookmarkService.java
@@ -9,10 +9,16 @@ import org.springframework.transaction.annotation.Transactional;
 
 import kr.taggle.bookmark.domain.Bookmark;
 import kr.taggle.bookmark.domain.BookmarkRepository;
+import kr.taggle.bookmark.domain.TagBookmark;
+import kr.taggle.bookmark.domain.TagBookmarkRepository;
 import kr.taggle.bookmark.dto.BookmarkCreateDto;
 import kr.taggle.bookmark.dto.BookmarkPageRequest;
 import kr.taggle.bookmark.dto.BookmarkResponse;
+import kr.taggle.bookmark.dto.BookmarkTagResponse;
+import kr.taggle.bookmark.dto.TagBookmarkResponse;
 import kr.taggle.bookmark.exception.BookmarkNotFoundException;
+import kr.taggle.tag.domain.Tag;
+import kr.taggle.tag.service.TagService;
 import kr.taggle.user.domain.User;
 import kr.taggle.user.dto.SessionUser;
 import kr.taggle.user.service.UserService;
@@ -25,6 +31,8 @@ public class BookmarkService {
 
     private final BookmarkRepository bookmarkRepository;
     private final UserService userService;
+    private final TagService tagService;
+    private final TagBookmarkRepository tagBookmarkRepository;
 
     public BookmarkResponse createBookmark(final SessionUser sessionUser, final BookmarkCreateDto bookmarkCreateDto) {
         final User user = userService.findById(sessionUser.getId());
@@ -41,6 +49,32 @@ public class BookmarkService {
 
         return BookmarkResponse.asList(bookmarks.getContent());
     }
+
+    //
+    @Transactional(readOnly = true)
+    public TagBookmarkResponse findBookmarksByTagId(final SessionUser user, final Long tagId,
+            final BookmarkPageRequest bookmarkPageRequest) {
+        final Tag tag = tagService.findByIdAndUserId(tagId, user.getId());
+        final Page<TagBookmark> bookmarks = tagBookmarkRepository.findAllByTag(tag, bookmarkPageRequest.toPageable());
+
+        return TagBookmarkResponse.of(tag, bookmarks.getContent());
+    }
+
+    @Transactional(readOnly = true)
+    public TagBookmarkResponse findUntaggedBookmarks(final SessionUser user,
+            final BookmarkPageRequest bookmarkPageRequest) {
+        final Page<Bookmark> bookmarks = findUntaggedBookmarksByUserId(user.getId(),
+                bookmarkPageRequest.toPageable());
+        return TagBookmarkResponse.ofUntagged(bookmarks.getContent());
+    }
+
+    @Transactional(readOnly = true)
+    public BookmarkTagResponse findTagsByBookmarkId(final SessionUser user, final Long id) {
+        final Bookmark bookmark = findByIdAndUserId(id, user.getId());
+
+        return BookmarkTagResponse.of(bookmark);
+    }
+    //
 
     public void removeBookmark(final SessionUser user, final Long id) {
         final Bookmark bookmark = findByIdAndUserId(id, user.getId());
