+++ backend/taggle-core/src/main/java/kr/taggle/bookmark/service/TagBookmarkService.java
@@ -1,15 +1,11 @@
 package kr.taggle.bookmark.service;
 
-import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import kr.taggle.bookmark.domain.Bookmark;
 import kr.taggle.bookmark.domain.TagBookmark;
 import kr.taggle.bookmark.domain.TagBookmarkRepository;
-import kr.taggle.bookmark.dto.BookmarkPageRequest;
-import kr.taggle.bookmark.dto.BookmarkTagResponse;
-import kr.taggle.bookmark.dto.TagBookmarkResponse;
 import kr.taggle.bookmark.exception.TagBookmarkNotFoundException;
 import kr.taggle.tag.domain.Tag;
 import kr.taggle.tag.service.TagService;
@@ -25,22 +21,29 @@ public class TagBookmarkService {
     private final BookmarkService bookmarkService;
     private final TagBookmarkRepository tagBookmarkRepository;
 
-    @Transactional(readOnly = true)
-    public TagBookmarkResponse findBookmarksByTagId(final SessionUser user, final Long tagId,
-            final BookmarkPageRequest bookmarkPageRequest) {
-        final Tag tag = tagService.findByIdAndUserId(tagId, user.getId());
-        final Page<TagBookmark> bookmarks = tagBookmarkRepository.findAllByTag(tag, bookmarkPageRequest.toPageable());
-
-        return TagBookmarkResponse.of(tag, bookmarks.getContent());
-    }
-
-    @Transactional(readOnly = true)
-    public TagBookmarkResponse findUntaggedBookmarks(final SessionUser user,
-            final BookmarkPageRequest bookmarkPageRequest) {
-        final Page<Bookmark> bookmarks = bookmarkService.findUntaggedBookmarksByUserId(user.getId(),
-                bookmarkPageRequest.toPageable());
-        return TagBookmarkResponse.ofUntagged(bookmarks.getContent());
-    }
+    // @Transactional(readOnly = true)
+    // public TagBookmarkResponse findBookmarksByTagId(final SessionUser user, final Long tagId,
+    //         final BookmarkPageRequest bookmarkPageRequest) {
+    //     final Tag tag = tagService.findByIdAndUserId(tagId, user.getId());
+    //     final Page<TagBookmark> bookmarks = tagBookmarkRepository.findAllByTag(tag, bookmarkPageRequest.toPageable());
+    //
+    //     return TagBookmarkResponse.of(tag, bookmarks.getContent());
+    // }
+    //
+    // @Transactional(readOnly = true)
+    // public TagBookmarkResponse findUntaggedBookmarks(final SessionUser user,
+    //         final BookmarkPageRequest bookmarkPageRequest) {
+    //     final Page<Bookmark> bookmarks = bookmarkService.findUntaggedBookmarksByUserId(user.getId(),
+    //             bookmarkPageRequest.toPageable());
+    //     return TagBookmarkResponse.ofUntagged(bookmarks.getContent());
+    // }
+    //
+    // @Transactional(readOnly = true)
+    // public BookmarkTagResponse findTagsByBookmarkId(final SessionUser user, final Long id) {
+    //     final Bookmark bookmark = bookmarkService.findByIdAndUserId(id, user.getId());
+    //
+    //     return BookmarkTagResponse.of(bookmark);
+    // }
 
     public void createTagBookmark(final SessionUser user, final Long bookmarkId, final Long tagId) {
         final Tag tag = tagService.findByIdAndUserId(tagId, user.getId());
@@ -60,12 +63,5 @@ public class TagBookmarkService {
 
         tagBookmarkRepository.delete(tagBookmark);
         bookmark.removeTagBookmark(tagBookmark);
-    }
-
-    @Transactional(readOnly = true)
-    public BookmarkTagResponse findTagsByBookmarkId(final SessionUser user, final Long id) {
-        final Bookmark bookmark = bookmarkService.findByIdAndUserId(id, user.getId());
-
-        return BookmarkTagResponse.of(bookmark);
     }
 }
