+++ backend/taggle-web/src/main/java/kr/taggle/bookmark/controller/BookmarkController.java
@@ -14,11 +14,14 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import kr.taggle.bookmark.dto.BookmarkCreateRequest;
 import kr.taggle.bookmark.dto.BookmarkPageRequest;
 import kr.taggle.bookmark.dto.BookmarkResponse;
+import kr.taggle.bookmark.dto.BookmarkTagResponse;
+import kr.taggle.bookmark.dto.TagBookmarkResponse;
 import kr.taggle.bookmark.service.BookmarkCreateService;
 import kr.taggle.bookmark.service.BookmarkService;
 import kr.taggle.user.dto.SessionUser;
@@ -48,6 +51,37 @@ public class BookmarkController {
 
         return ResponseEntity.ok()
                 .body(bookmarks);
+    }
+
+    @GetMapping("/{bookmarkId}")
+    public ResponseEntity<BookmarkTagResponse> findBookmarkDetail(@AuthenticationPrincipal final SessionUser user,
+            @PathVariable final Long bookmarkId) {
+        final BookmarkTagResponse bookmarkTagResponse = bookmarkService.findTagsByBookmarkId(user, bookmarkId);
+
+        return ResponseEntity.ok()
+                .body(bookmarkTagResponse);
+    }
+
+    @GetMapping(params = "tag")
+    public ResponseEntity<TagBookmarkResponse> findBookmarksByTagId(@AuthenticationPrincipal final SessionUser user,
+            @RequestParam final Long tag,
+            @ModelAttribute final BookmarkPageRequest bookmarkPageRequest) {
+        final TagBookmarkResponse tagBookmarkResponse = bookmarkService.findBookmarksByTagId(user, tag,
+                bookmarkPageRequest);
+
+        return ResponseEntity.ok()
+                .body(tagBookmarkResponse);
+    }
+
+    @GetMapping(params = "tag=none")
+    public ResponseEntity<TagBookmarkResponse> findBookmarksWithUntagged(
+            @AuthenticationPrincipal final SessionUser user,
+            @ModelAttribute final BookmarkPageRequest bookmarkPageRequest) {
+        final TagBookmarkResponse tagBookmarkResponse = bookmarkService.findUntaggedBookmarks(user,
+                bookmarkPageRequest);
+
+        return ResponseEntity.ok()
+                .body(tagBookmarkResponse);
     }
 
     @DeleteMapping("/{id}")
